+++ Changelog.md
@@ -7,11 +7,13 @@
 * changed some things about the API, it's possible your code will break sorry
 * fixed some bugs
 * if you supply a `defaultFileText` when creating a `DataFile`, if you use `Get` and you don't provide a default value, the default value will be searched for in the `defaultFileText`
-* added `DistributedData` class
+* added `DistributedData` class, and many extension methods for it
 * added `DataFile.TryGet`
   * todo: document this, add non-generic version, add path based versions
 * added ReadableWritableDataFile.ResetToDefaultData()
 * fixed default data path on Mac OS (Unity version)
+* added `ReadableDataFile.Identifier`
+  * when creating a `MemoryDataFile`, you can now supply an identifier
 
 todo update the docs for 1.2